requirements.txt
fastapi~=0.115.4
python-dotenv~=1.0.1
SQLAlchemy~=2.0.36
PyMySQL~=1.1.1
pydantic~=2.9.2
h11~=0.14.0
pip~=24.3.1
typing_extensions~=4.12.2
sniffio~=1.3.1
click~=8.1.7
starlette~=0.41.2
pydantic_core~=2.23.4
anyio~=4.6.2.post1
uvicorn~=0.32.0